refactor: use the term "branch planner" everywhere
Closes #705

References:
* https://github.com/weaveworks/tf-controller/issues/705
