+++ Tiltfile
@@ -1,7 +1,7 @@
 local('kubectl apply --server-side -k config/tilt/base')
 
 k8s_yaml(kustomize('config/tilt/manager'))
-k8s_yaml(kustomize('config/tilt/bbp'))
+k8s_yaml(kustomize('config/tilt/branch-planner'))
 
 docker_build(
     'weaveworks/tf-controller',
@@ -16,7 +16,7 @@ custom_build(
 )
 
 docker_build(
-    'weaveworks/branch-based-planner',
+    'weaveworks/branch-planner',
     context='.',
     dockerfile='planner.Dockerfile',
 )
