+++ config/tilt/branch-planner/kustomization.yaml
@@ -2,7 +2,7 @@ apiVersion: kustomize.config.k8s.io/v1beta1
 kind: Kustomization
 namespace: flux-system
 resources:
-- ../../bbp
+- ../../branch-planner
 images:
-- name: ghcr.io/weaveworks/branch-based-planner
-  newName: weaveworks/branch-based-planner
+- name: ghcr.io/weaveworks/branch-planner
+  newName: weaveworks/branch-planner
