+++ docs/References/terraform.md
@@ -1643,13 +1643,7 @@ string
 </td>
 <td>
 <em>(Optional)</em>
-<p>StoreReadablePlan enables storing the plan in a readable format. Defaults to 'none'.</p>
-<p>This is an Enum and has the expected values of:</p>
-<ul>
-<li>human</li>
-<li>json</li>
-<li>none</li>
-</ul>
+<p>StoreReadablePlan enables storing the plan in a readable format.</p>
 </td>
 </tr>
 <tr>
@@ -2171,13 +2165,7 @@ string
 </td>
 <td>
 <em>(Optional)</em>
-<p>StoreReadablePlan enables storing the plan in a readable format. Defaults to 'none'.</p>
-<p>This is an Enum and has the expected values of:</p>
-<ul>
-<li>human</li>
-<li>json</li>
-<li>none</li>
-</ul>
+<p>StoreReadablePlan enables storing the plan in a readable format.</p>
 </td>
 </tr>
 <tr>
